[MOVED] `WeatherService` to data model module
