+++ app/src/main/java/dev/hossain/weatheralert/data/ApiKeyProvider.kt
@@ -2,6 +2,7 @@ package dev.hossain.weatheralert.data
 
 import com.squareup.anvil.annotations.ContributesBinding
 import dev.hossain.weatheralert.BuildConfig
+import dev.hossain.weatheralert.datamodel.WeatherService
 import dev.hossain.weatheralert.di.AppScope
 import javax.inject.Inject
