+++ app/src/main/java/dev/hossain/weatheralert/data/WeatherRepository.kt
@@ -8,6 +8,7 @@ import dev.hossain.weatheralert.datamodel.AppForecastData
 import dev.hossain.weatheralert.datamodel.Rain
 import dev.hossain.weatheralert.datamodel.Snow
 import dev.hossain.weatheralert.datamodel.WeatherApiServiceResponse
+import dev.hossain.weatheralert.datamodel.WeatherService
 import dev.hossain.weatheralert.db.CityForecast
 import dev.hossain.weatheralert.db.CityForecastDao
 import dev.hossain.weatheralert.di.AppScope
