+++ app/src/main/java/dev/hossain/weatheralert/work/WeatherCheckWorker.kt
@@ -9,8 +9,8 @@ import dagger.assisted.Assisted
 import dagger.assisted.AssistedInject
 import dev.hossain.weatheralert.data.PreferencesManager
 import dev.hossain.weatheralert.data.WeatherRepository
-import dev.hossain.weatheralert.data.WeatherService
 import dev.hossain.weatheralert.datamodel.WeatherAlertCategory
+import dev.hossain.weatheralert.datamodel.WeatherService
 import dev.hossain.weatheralert.db.AlertDao
 import dev.hossain.weatheralert.db.UserCityAlert
 import dev.hossain.weatheralert.notification.triggerNotification
